Deleting blogs using event handler throught props
An event onclick handler sends the blog id that will be deleted.

Using a filter function, the handleDelete() funciton saves all the
blogs except that one that has the id blog received, in a variable
called 'newBlogs'.

The setBlog() function uses the newBlogs variable to update the blogs
object without destroying the original one.